+++ runtime/src/flow_job/mod.rs
@@ -2,10 +2,12 @@ mod block_request;
 pub mod flow;
 mod node_input_values;
 mod run_to_node;
-
+mod upstream;
 pub use block_request::{
     parse_node_downstream, parse_query_block_request, parse_run_block_request,
     RunBlockSuccessResponse,
 };
-pub use flow::{execute_flow_job, find_upstream, FlowJobParameters, UpstreamArgs};
+pub use flow::{execute_flow_job, FlowJobParameters};
 pub use node_input_values::NodeInputValues;
+pub(crate) use upstream::find_upstream_nodes;
+pub use upstream::{find_upstream, UpstreamParameters};
